+++ conftest.py
@@ -5,15 +5,15 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-from postgresql_audit import read_file
+from postgresql_audit import versioning_manager
 
 
-@pytest.fixture(scope='module')
+@pytest.fixture()
 def dns():
     return 'postgres://postgres@localhost/postgresql_audit_test'
 
 
-@pytest.fixture(scope='module')
+@pytest.fixture()
 def base():
     return declarative_base()
 
@@ -29,8 +29,8 @@ def engine(dns):
 @pytest.yield_fixture()
 def connection(engine):
     conn = engine.connect()
-    conn.execute('DROP SCHEMA IF EXISTS audit CASCADE')
     conn.execute('CREATE EXTENSION IF NOT EXISTS hstore')
+    conn.execute('DROP SCHEMA IF EXISTS audit CASCADE')
     yield conn
     conn.close()
 
@@ -44,58 +44,48 @@ def session(connection):
     session.close_all()
 
 
-@pytest.yield_fixture()
-def schema(session):
-    files = [
-        'schema.sql',
-        'activity.sql',
-        'create_activity.sql',
-        'audit_table.sql'
-    ]
-    for file_ in files:
-        session.execute(read_file(file_))
-    session.commit()
-    yield
-    session.execute('DROP SCHEMA audit CASCADE')
-    session.commit()
+@pytest.fixture()
+def activity_cls(base):
+    versioning_manager.init(base)
+    return versioning_manager.activity_cls
 
 
-@pytest.fixture(scope='module')
+@pytest.fixture()
 def user_class(base):
     class User(base):
         __tablename__ = 'user'
+        __versioned__ = {}
         id = sa.Column(sa.Integer, primary_key=True)
         name = sa.Column(sa.String(100))
         age = sa.Column(sa.Integer)
     return User
 
 
-@pytest.fixture(scope='module')
+@pytest.fixture()
 def article_class(base):
     class Article(base):
         __tablename__ = 'article'
+        __versioned__ = {}
         id = sa.Column(sa.Integer, primary_key=True)
         name = sa.Column(sa.String(100))
     return Article
 
 
-@pytest.fixture(scope='module')
+@pytest.fixture()
 def models(user_class, article_class):
     return [user_class, article_class]
 
 
 @pytest.yield_fixture
-def table_creator(base, connection, session, models):
+def table_creator(base, connection, session, models, activity_cls):
     sa.orm.configure_mappers()
-    base.metadata.create_all(connection, checkfirst=True)
     tx = connection.begin()
-    for model in models:
-        connection.execute(
-            "SELECT audit.audit_table('{0}')".format(model.__tablename__)
-        )
+    versioning_manager.activity_cls.__table__.create(connection)
+    base.metadata.create_all(connection)
     tx.commit()
     yield
     base.metadata.drop_all(connection)
+    session.commit()
 
 
 @pytest.fixture