+++ postgresql_audit/__init__.py
@@ -4,10 +4,10 @@ from contextlib import contextmanager
 from weakref import WeakSet
 
 import sqlalchemy as sa
-from sqlalchemy.dialects.postgresql import HSTORE, INET
+from sqlalchemy.dialects.postgresql import array, HSTORE, INET
 from sqlalchemy.ext.compiler import compiles
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.schema import CreateTable
+from sqlalchemy.schema import CreateTable, DropTable
 
 
 __version__ = '0.1.7'
@@ -19,7 +19,7 @@ class ImproperlyConfigured(Exception):
 
 
 @compiles(CreateTable)
-def _add_if_not_exists(element, compiler, **kw):
+def _add_if_not_exists_to_create(element, compiler, **kw):
     output = compiler.visit_create_table(element, **kw)
     if element.element.info.get('ifexists'):
         output = re.sub(
@@ -31,10 +31,45 @@ def _add_if_not_exists(element, compiler, **kw):
     return output
 
 
+@compiles(DropTable)
+def _add_if_not_exists_to_drop(element, compiler, **kw):
+    output = compiler.visit_drop_table(element, **kw)
+    if element.element.info.get('ifexists'):
+        output = re.sub(
+            '^\s*(DROP (TEMP )?TABLE)',
+            '\\1 IF EXISTS',
+            output,
+            re.DOTALL
+        )
+    return output
+
+
+class StatementExecutor(object):
+    def __init__(self, stmt):
+        self.stmt = stmt
+
+    def __call__(self, target, bind, **kwargs):
+        bind.execute(self.stmt)
+
+
+def get_cursor(conn):
+    if isinstance(conn, sa.engine.Engine):
+        return conn.raw_connection().connection.cursor()
+    elif isinstance(conn, sa.engine.Connection):
+        return conn.connection.cursor()
+    else:
+        return conn.cursor()
+
+
 def raw_execute(conn, stmt):
-    cursor = conn.connection.cursor()
-    compiled = stmt.compile(conn.engine)
-    return cursor.execute(str(compiled), compiled.params)
+    cursor = get_cursor(conn)
+    if isinstance(stmt, str):
+        compiled = stmt
+        params = []
+    else:
+        compiled = stmt.compile(conn.engine)
+        params = compiled.params
+    return cursor.execute(str(compiled), params)
 
 
 def read_file(file_):
@@ -93,6 +128,7 @@ def activity_base(base, actor_cls):
 
 @contextmanager
 def activity_values(conn, **kwargs):
+    versioning_manager.create_temp_table_if_not_exists(conn)
     versioning_manager.set_activity_values(conn, **kwargs)
     yield
     versioning_manager.reset_activity_values(conn)
@@ -123,16 +159,69 @@ class VersioningManager(object):
         self._actor_cls = actor_cls
         self.connections_with_tables = WeakSet()
         self.connections_with_tables_row = WeakSet()
-        self.pool_listener_args = (
-            sa.pool.Pool,
-            'checkin',
-            self.connection_listener
+        self.listeners = (
+            (
+                sa.pool.Pool,
+                'checkin',
+                self.connection_listener
+            ),
+            (
+                sa.engine.Engine,
+                'before_cursor_execute',
+                self.before_cursor_execute
+            ),
+            (
+                sa.orm.mapper,
+                'instrument_class',
+                self.instrument_versioned_classes
+            ),
+            (
+                sa.orm.mapper,
+                'after_configured',
+                self.configure_versioned_classes
+            )
         )
-        self.engine_listener_args = (
-            sa.engine.Engine,
-            'before_cursor_execute',
-            self.before_cursor_execute
-        )
+        self.pending_classes = WeakSet()
+        self.cached_ddls = {}
+
+    def instrument_versioned_classes(self, mapper, cls):
+        """
+        Collect versioned class and add it to pending_classes list.
+
+        :mapper mapper: SQLAlchemy mapper object
+        :cls cls: SQLAlchemy declarative class
+        """
+        if hasattr(cls, '__versioned__') and cls not in self.pending_classes:
+            self.pending_classes.add(cls)
+
+    def configure_versioned_classes(self):
+        """
+        Configures all versioned classes that were collected during
+        instrumentation process.
+        """
+        for cls in self.pending_classes:
+            table = cls.__table__
+            args = [table.name]
+            exclude = cls.__versioned__.get('exclude')
+            if exclude:
+                for column in exclude:
+                    if column not in table.c:
+                        raise ImproperlyConfigured(
+                            "Could not configure versioning. Table '{}'' does "
+                            "not have a column named '{}'.".format(
+                                table.name, column
+                            )
+                        )
+                args.append(array(exclude))
+
+            query = sa.select(
+                [sa.func.audit.audit_table(*args)]
+            )
+            if query not in self.cached_ddls:
+                self.cached_ddls[query] = StatementExecutor(query)
+            listener = (table, 'after_create', self.cached_ddls[query])
+            if not sa.event.contains(*listener):
+                sa.event.listen(*listener)
 
     def attach_table_listeners(self):
         for values in self.table_listeners:
@@ -158,20 +247,29 @@ class VersioningManager(object):
         # Use raw cursor so that SQLAlchemy events are not invoked
         raw_execute(conn, stmt.values(**kwargs_copy))
 
-    def reset_activity_values(self, conn):
-        raw_execute(conn, self.table.delete())
-        self.connections_with_tables_row.remove(conn.connection.connection)
-
-    def connection_listener(self, dbapi_connection, connection_record):
+    def reset_connection_cache(self, dbapi_connection):
         self.connections_with_tables.discard(dbapi_connection)
         self.connections_with_tables_row.discard(dbapi_connection)
 
-    def create_temp_table(self, cursor, conn):
-        cursor.execute(
+    def reset_activity_values(self, conn):
+        raw_execute(conn, DropTable(self.table))
+        self.reset_connection_cache(conn.connection.connection)
+
+    def connection_listener(self, dbapi_connection, connection_record):
+        self.reset_connection_cache(dbapi_connection)
+
+    def create_temp_table(self, conn):
+        conn.execute(
             str(CreateTable(self.table).compile(conn.engine))
             +
             ' ON COMMIT PRESERVE ROWS'
         )
+
+    def create_temp_table_if_not_exists(self, conn):
+        raw_conn = conn.connection.connection
+        if raw_conn not in self.connections_with_tables:
+            self.create_temp_table(conn)
+            self.connections_with_tables.add(raw_conn)
 
     def before_cursor_execute(
         self,
@@ -182,12 +280,11 @@ class VersioningManager(object):
         context,
         executemany
     ):
-        raw_conn = conn.connection.connection
-        if raw_conn not in self.connections_with_tables:
-            self.create_temp_table(cursor, conn)
-            self.connections_with_tables.add(raw_conn)
-        if self.values:
-            self.set_activity_values(conn, **self.values)
+        stmt = statement.strip().lower()
+        if stmt.startswith(('insert', 'update', 'delete')):
+            self.create_temp_table_if_not_exists(conn)
+            if self.values:
+                self.set_activity_values(conn, **self.values)
 
     @property
     def actor_cls(self):
@@ -215,13 +312,13 @@ class VersioningManager(object):
 
     def attach_listeners(self):
         self.attach_table_listeners()
-        sa.event.listen(*self.engine_listener_args)
-        sa.event.listen(*self.pool_listener_args)
+        for listener in self.listeners:
+            sa.event.listen(*listener)
 
     def remove_listeners(self):
         self.remove_table_listeners()
-        sa.event.remove(*self.engine_listener_args)
-        sa.event.remove(*self.pool_listener_args)
+        for listener in self.listeners:
+            sa.event.remove(*listener)
 
     def activity_model_factory(self, base):
         class Activity(activity_base(base, self.actor_cls)):